add:auth画面のリファクタリングをしました
