+++ lib/mixin/account_exist_mixin.dart
@@ -1,6 +1,5 @@
 import 'dart:async';
 import 'package:food_gram_app/main.dart';
-import 'package:supabase_flutter/supabase_flutter.dart';
 
 mixin AccountExistMixin {
   Future<bool> doesAccountExist() async {
@@ -10,22 +9,9 @@ mixin AccountExistMixin {
     }
     final response =
         await supabase.from('users').select().eq('user_id', user.id).execute();
-
     if (response.data != null && response.data.isNotEmpty) {
       return true;
     }
     return false;
-  }
-
-  Future<bool> loginStatus() async {
-    final completer = Completer<bool>();
-    Supabase.instance.client.auth.onAuthStateChange.listen((state) {
-      if (state.event == AuthChangeEvent.signedIn) {
-        completer.complete(true);
-      } else {
-        completer.complete(false);
-      }
-    });
-    return completer.future;
   }
 }
